requirements.txt
#
# Copyright 2022 The AI Flow Authors
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
#
# Required dependencies
streamz==0.5.2
matplotlib==3.1.2
grpcio==1.35.0
requests==2.22.0
six==1.12.0
kafka-python==2.0.2
kubernetes==9.0.0
networkx==2.4
pika==1.1.0
protobuf==3.15.6
pyyaml>=5.1, <5.4
apscheduler==3.6.3
google-api-core==1.16.0
psutil==5.6.3
pika==1.1.0
pyecharts==1.8.1
cloudpickle==1.2.2
flask>=1.1.0, <2.0
flask-admin==1.5.4
flask-appbuilder>=1.12.2, <2.0.0;python_version<"3.6"
flask-appbuilder~=3.1.1;python_version>="3.6"
flask-caching~=1.9.0
flask_cors>=3.0.10
flask-login>=0.3, <0.5
flask-swagger==0.2.13
flask-wtf>=0.14.2, <0.15
Django~=3.2.7
WTForms==2.3.3
mongoengine~=0.22.1
python-daemon>=2.1.1, <2.2
MarkupSafe==1.1.1

# Optional dependencies
alembic>=1.0,<2.0
argcomplete~=1.10
attrs~=20.3.0
cached_property~=1.5
colorlog==4.0.2
configparser>=3.5.0,<3.6.0
croniter>=0.3.17, <0.4
dill>=0.2.2, <0.4
enum34~=1.1.6;python_version<"3.4"
funcsigs>=1.0.0, <2.0.0
future>=0.16.0, <0.19
graphviz>=0.12
gunicorn>=19.5.0, <20.0
iso8601>=0.1.12
jinja2>=2.10.1, <2.11.0
json-merge-patch==0.2
jsonschema~=3.0
lazy_object_proxy~=1.3
markdown>=2.5.2, <3.0
marshmallow-sqlalchemy>=0.16.1, <0.19.0;python_version<"3.6"
pendulum~=2.0
pygments>=2.0.1, <3.0
python-dateutil>=2.3, <3
setproctitle>=1.1.8, <2
sqlalchemy==1.3.23
sqlalchemy_jsonfield==0.8.0;python_version<"3.5"
sqlalchemy_jsonfield~=1.0;python_version>="3.5"
tabulate>=0.7.5, <0.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
thrift>=0.9.2
typing;python_version<"3.5"
typing-extensions>=3.7.4;python_version<"3.8"
tzlocal>=1.4,<2.0.0
unicodecsv>=0.14.1
werkzeug~=1.0
zope.deprecation>=4.0, <5.0
